// src/app/core/enums/install-type.ts
export enum InstallType {
  UNSPECIFIED = 'INSTALL_TYPE_UNSPECIFIED',
  PREINSTALLED = 'PREINSTALLED',
  FORCE_INSTALLED = 'FORCE_INSTALLED',
  BLOCKED = 'BLOCKED',
  AVAILABLE = 'AVAILABLE',
  REQUIRED_FOR_SETUP = 'REQUIRED_FOR_SETUP',
  KIOSK = 'KIOSK',
}

export enum InstallTypeDescription {
  UNSPECIFIED = 'Sin especificar',
  PREINSTALLED = 'Preinstalada',
  FORCE_INSTALLED = 'Forzar instalación',
  BLOCKED = 'Bloqueada',
  AVAILABLE = 'Disponible',
  REQUIRED_FOR_SETUP = 'Necesaria para la configuración',
  KIOSK = 'Modo kiosko',
}