// src/app/core/enums/permission-policy.ts
export enum PermissionPolicy {
  UNSPECIFIED = 'PERMISSION_POLICY_UNSPECIFIED',
  PROMPT = 'PROMPT',
  GRANT = 'GRANT',
  DENY = 'DENY',
}

export enum PermissionPolicyDescription {
  UNSPECIFIED = 'Sin especificar',
  PROMPT = 'Preguntar',
  GRANT = 'Otorgar permisos',
  DENY = 'Denegar permisos',
}