// src/app/core/enums/auto-update-mode.ts
export enum AutoUpdateMode {
  UNSPECIFIED = 'AUTO_UPDATE_MODE_UNSPECIFIED',
  DEFAULT = 'AUTO_UPDATE_DEFAULT',
  POSTPONED = 'AUTO_UPDATE_POSTPONED',
  HIGH_PRIORITY = 'AUTO_UPDATE_HIGH_PRIORITY',
}

export enum AutoUpdateModeDescription {
  UNSPECIFIED = 'Sin especificar',
  DEFAULT = 'Por defecto',
  POSTPONED = 'Posponer',
  HIGH_PRIORITY = 'Prioritario',
}