// src/app/core/enums/wifi-state.ts
export enum WifiState {
  UNSPECIFIED = 'WIFI_STATE_UNSPECIFIED',
  USER_CHOICE = 'WIFI_STATE_USER_CHOICE',
  ENABLED = 'WIFI_ENABLED',
  DISABLED = 'WIFI_DISABLED',
}

export enum WifiStateDescription {
  UNSPECIFIED = 'Sin especificar',
  USER_CHOICE = 'Permitir elegir al usuario',
  ENABLED = 'Activado',
  DISABLED = 'Desactivado',
}