// src/app/core/enums/location-mode.ts
export enum LocationMode {
  UNSPECIFIED = 'LOCATION_MODE_UNSPECIFIED',
  USER_CHOICE = 'LOCATION_USER_CHOICE',
  ENFORCED = 'LOCATION_ENFORCED',
  DISABLED = 'LOCATION_DISABLED',
}

export enum LocationModeDescription {
  UNSPECIFIED = 'Sin especificar',
  USER_CHOICE = 'Permitir elegir al usuario',
  ENFORCED = 'Activada',
  DISABLED = 'Desactivada',
}