// src/app/core/enums/usb-data-access.ts
export enum UsbDataAccess {
  UNSPECIFIED = 'USB_DATA_ACCESS_UNSPECIFIED',
  ALLOW_USB_DATA_TRANSFER = 'ALLOW_USB_DATA_TRANSFER',
  DISALLOW_USB_FILE_TRANSFER = 'DISALLOW_USB_FILE_TRANSFER',
  DISALLOW_USB_DATA_TRANSFER = 'DISALLOW_USB_DATA_TRANSFER',
}

export enum UsbDataAccessDescription {
  UNSPECIFIED = 'Sin especificar',
  ALLOW_USB_DATA_TRANSFER = 'Permitir transefencia de datos',
  DISALLOW_USB_FILE_TRANSFER = 'No permitir transefencia de archivos',
  DISALLOW_USB_DATA_TRANSFER = 'No permitir transefencia de datos',
}