// src/app/core/enums/wifi-ssid-policy-type.ts
export enum WifiSsidPolicyType {
  UNSPECIFIED = 'WIFI_SSID_POLICY_TYPE_UNSPECIFIED',
  DENYLIST = 'WIFI_SSID_DENYLIST',
  ALLOWLIST = 'WIFI_SSID_ALLOWLIST',
}

export enum WifiSsidPolicyTypeDescription {
  UNSPECIFIED = 'Sin especificar',
  DENYLIST = 'Lista de bloqueados',
  ALLOWLIST = 'Lista de permitidos',
}