// src/app/core/enums/developer-settings.ts
export enum DeveloperSettings {
  UNSPECIFIED = 'DEVELOPER_SETTINGS_UNSPECIFIED',
  DISABLED = 'DEVELOPER_SETTINGS_DISABLED',
  ALLOWED = 'DEVELOPER_SETTINGS_ALLOWED',
}

export enum DeveloperSettingsDescription {
  UNSPECIFIED = 'Sin especificar',
  DISABLED = 'Desactivado',
  ALLOWED = 'Activado',
}