// src/app/core/enums/screen-brightness-mode.ts
export enum ScreenBrightnessMode {
  UNSPECIFIED = 'SCREEN_BRIGHTNESS_MODE_UNSPECIFIED',
  USER_CHOICE = 'BRIGHTNESS_USER_CHOICE',
  AUTOMATIC = 'BRIGHTNESS_AUTOMATIC',
  FIXED = 'BRIGHTNESS_FIXED',
}

export enum ScreenBrightnessModeDescription {
  UNSPECIFIED = 'Sin especificar',
  USER_CHOICE = 'Permitir elegir al usuario',
  AUTOMATIC = 'Automático',
  FIXED = 'Brillo fijo',
}