// src/app/core/enums/screen-timeout-mode.ts
export enum ScreenTimeoutMode {
  UNSPECIFIED = 'SCREEN_TIMEOUT_MODE_UNSPECIFIED',
  USER_CHOICE = 'SCREEN_TIMEOUT_USER_CHOICE',
  ENFORCED = 'SCREEN_TIMEOUT_ENFORCED',
}

export enum ScreenTimeoutModeDescription {
  UNSPECIFIED = 'Sin especificar',
  USER_CHOICE = 'Permitir elegir al usuario',
  ENFORCED = 'Activado',
}